fix bug with price in order

+++ src/main/java/serwisPaczek/model/UserOrder.java
@@ -14,7 +14,7 @@ public class UserOrder {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-    private float price;
+    private double price;
     private Date date;
     private int premiumPoints;
     @ManyToOne
@@ -39,7 +39,7 @@ public class UserOrder {
     public UserOrder() {
     }
 
-    public UserOrder(float price, Date date, User user, Courier courier, Status status, SenderAdress senderAdress, RecipientAdress recipientAdress, Parcel parcel) {
+    public UserOrder(double price, Date date, User user, Courier courier, Status status, SenderAdress senderAdress, RecipientAdress recipientAdress, Parcel parcel) {
         this.price = price;
         this.date = date;
         this.user = user;
@@ -50,7 +50,8 @@ public class UserOrder {
         this.parcel = parcel;
     }
 
-    public UserOrder(float price, Date date, int premiumPoints, User user, Courier courier, Status status, SenderAdress senderAdress, RecipientAdress recipientAdress, Parcel parcel) {
+    public UserOrder(double price, Date date, int premiumPoints, User user, Courier courier, Status status,
+                     SenderAdress senderAdress, RecipientAdress recipientAdress, Parcel parcel) {
         this.price = price;
         this.date = date;
         this.premiumPoints = premiumPoints;
@@ -62,7 +63,8 @@ public class UserOrder {
         this.parcel = parcel;
     }
 
-    public UserOrder(float price, Date date, User user, Courier courier, Status status, SenderAdress senderAdress, RecipientAdress recipientAdress) {
+    public UserOrder(double price, Date date, User user, Courier courier, Status status, SenderAdress senderAdress,
+                     RecipientAdress recipientAdress) {
         this.price = price;
         this.date = date;
         this.user = user;
